+++ src/components/useSiteRegisterDialog.ts
@@ -1,8 +1,6 @@
-import { SitePostRequest } from './../types/request/site.ts'
-import { SitePostResponse } from './../types/response/site.ts'
-// eslint-disable-next-line import/named
-import axios, { AxiosResponse } from 'axios'
 import { useState } from 'react'
+import postSite from '../useCases/postSiteUseCase.ts'
+import postSiteCheck from '../useCases/postSiteCheck.ts'
 
 const useSiteRegisterDialog = () => {
   const [name, setName] = useState<string>('')
@@ -12,48 +10,26 @@ const useSiteRegisterDialog = () => {
   const [metaTag, setMetaTag] = useState<string>('')
   const [step, setStep] = useState<number>(1)
 
-  const onClickNextButton = () => {
-    axios
-      .post('/sites', {
-        name,
-        url,
-        description,
-      } as SitePostRequest)
-      .then((response: AxiosResponse<SitePostResponse>) => {
-        console.log(response)
-        setMetaTag(response.data.metaTag)
-        setStep(2)
-        return response.data
-      })
-      .catch(function (error) {
-        console.log(error)
-        //TODO : remove this dummy code
-        setMetaTag('TESTDUMMY234234')
-        setStep(2)
-      })
-  }
-
-  const onClickCheckButton = () => {
-    axios
-      .post('/check-sites', {})
-      .then((response) => {
-        console.log(response)
-        setStep(3)
-        return response.data
-      })
-      .catch(function (error) {
-        console.log(error)
-        //TODO : remove this dummy code
-        setStep(4)
-      })
-  }
-
-  const copyMetaTag = async () => {
+  const onClickNextButton = async () => {
     try {
-      await navigator.clipboard.writeText(`<meta name="quack-run-site-verification" content="${copyMetaTag}">`)
-    } catch (e) {
-      console.log(e)
+      const res = await postSite({ name, url, description })
+      setMetaTag(res.metaTag)
+      setStep(2)
+    } catch {
+    //TODO: 안내 토스트
     }
+  }
+
+  const onClickCheckButton = async () => {
+    postSiteCheck({ id: 1 })
+    //성공
+    setStep(3)
+    //실패
+    setStep(4)
+  }
+
+  const copyMetaTag = () => {
+    navigator.clipboard.writeText(`<meta name="quack-run-site-verification" content="${copyMetaTag}">`)
   }
 
   const onCloseModal = () => {
